refactor: 프래그먼트에서 일관성 없는 view nullable 변경
- 모든 프래그먼트에서 view가 null을 가지지 않도록 수정
- 그 외에도 receiver의 패키지, intent key 상수화 등 수정
