+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
 
-        <receiver android:name=".ui.confirm.AlarmReceiver" />
+        <receiver android:name=".ui.receiver.AlarmReceiver" />
 
         <activity
             android:name=".ui.adv.AdvDetailActivity"
