+++ app/src/main/java/woowacourse/movie/ui/receiver/AlarmReceiver.kt
@@ -1,33 +1,33 @@
-package woowacourse.movie.ui.confirm
+package woowacourse.movie.ui.receiver
 
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.util.Log
-import woowacourse.movie.util.PreferenceUtil
 import woowacourse.movie.R
 import woowacourse.movie.model.TicketsState
 import woowacourse.movie.ui.fragment.setting.SettingFragment.Companion.NOTIFICATIONS
+import woowacourse.movie.util.PreferenceUtil
 import woowacourse.movie.util.getParcelableCompat
 import woowacourse.movie.util.sendNotification
 
 class AlarmReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context, intent: Intent?) {
-        val tickets = intent?.extras?.getParcelableCompat<TicketsState>("a")
-
+        val tickets: TicketsState = intent?.extras?.getParcelableCompat(KEY_TICKETS) ?: return
         val sharedPreference = PreferenceUtil(context)
         val isNotification = sharedPreference.getBoolean(NOTIFICATIONS, false)
 
         if (isNotification) {
-            tickets?.let {
-                context.sendNotification(
-                    R.drawable.ic_launcher_foreground,
-                    context.getString(R.string.alarm_receiver_notification_title),
-                    context.getString(R.string.alarm_receiver_notification_text)
-                        .format(it.movieState.title),
-                    true
-                )
-            }
+            context.sendNotification(
+                R.drawable.ic_launcher_foreground,
+                context.getString(R.string.alarm_receiver_notification_title),
+                context.getString(R.string.alarm_receiver_notification_text)
+                    .format(tickets.movieState.title),
+                true
+            )
         }
+    }
+
+    companion object {
+        private const val KEY_TICKETS = "key_tickets"
     }
 }
